Adding structure for Tracing & Timing
This is an experimentation where when launching a new chapter,
StoryTeller would trace the chapter also run a timing mechanism to know
how long a chapter would run.

The Timer right now is pretty naive and only time the duration of the
whole block. It's not possible right now to have a breakdown of how long
each call processed. That could be useful when hitting multiple
datastore (sql, redis, etc.) in a single chapter to isolate which one of
the call was the more costly. I'll need to investigate whether or not
this is possible within the scope of a chapter