Encoding: Gracefully handle bad encoding
Because we end up parsing it to JSON and all the string needs to end up
in the same encoding, the template is now forced to be UTF-8. I'm not
sure how this will play out, however because maybe UTF-8 is not as
ubiquitous as I think it is.

Also, in my experience, non-utf8 encoding usually is based on bad
requests/data and not on expected use case. Most browsers today defaults
to UTF8. Maybe asian browser uses a different encoding. Back in the
days, Japan had its own encoding that was basically an ASCII version for
their dialect. In that case, this wouldn't work and we'd need to change
the behavior.

In the end, the reason for these checks is because I've seen weird
string representation in Rails where we have some odd UTF-8 character.
Messages shouldn't be affected by that, because these are hard coded
string by the developers. But I was creating spec and thought it could
possibly happen, so here it is.