Chapter only raise error if they're the first to notice
Previous implementation would have created an error story for each of
the chapter that saw the error. By storing the last uncaught error in
the current chapter after popping, it allows us to recover error within
another chapter and also catching any other error that would be
triggered.

That makes it for a cleaner API that will only log exception once while
keeping the same behavior of catching errors inside chapters as before.