Remove message requirement
The main app has occurences where message are nil and so I don't want to
introduce a regression over there. Instead of forcing everyone to change
the code, I'm just making some changes to the code so StoryTeller can be
backward compatible with the structure we already have in place on the
main app.

It's possible in the future it is reintroduce, but that will have to be
discussed with every users of ST.